pcxt: move shared DS layout out of the page pool
Reserve the rest of the first 64 KiB segment for shared DS=0
state and move the first free page to 0x10000. This fixes the
pcxt bring-up layout so mem_region pages no longer overlap the
shared core/VFS area.

Update the ia16 page-pool base handling to use 32-bit linear
addresses, add a VFS linker assertion for the DS=0 window, and
report both ELF size and segment usage during pcxt builds.
The headless run path now keeps COM1 on stdio, and the PIT start
is deferred until waitpid blocks so early floppy-backed startup
work can complete before timer IRQs begin.

Update the port proposal to reflect current progress: headless
serial and BIOS tty output are fixed, the old /etc/inittab read
failure is gone, and the remaining blocker is the post-init
handoff after init starts.

Verification:
- ./scripts/build.sh pcxt
- ./scripts/run.sh pcxt
  Boot now reaches:
  INIT: starting
  INIT: /sbin/init failed, trying /bin/sh

+++ src/kernel/mm/page.c
@@ -154,7 +154,7 @@ void mm_init(void) {
    * with the kernel stack.  On QEMU (flat memory model) the initial
    * stack may extend into the page pool region.  Skip those pages. */
   uintptr_t stack_page_top = (stack_top + PAGE_SIZE - 1u) & ~(PAGE_SIZE - 1u);
-  uintptr_t pool_base;
+  uint32_t pool_base;
 #if defined(__xtensa__)
   pool_base = xtensa_pool_base;
 #else
@@ -254,7 +254,7 @@ void *page_alloc_at(void *addr) {
   uintptr_t target = (uintptr_t)addr;
 
   /* Validate: must be page-aligned and within the runtime pool */
-  uintptr_t pb = page_pool_base();
+  uint32_t pb = page_pool_base();
   if (target < pb || target >= pb + page_count * PAGE_SIZE)
     return NULL;
   if (target & (PAGE_SIZE - 1u)) return NULL;
@@ -310,7 +310,7 @@ static void stack_backtrace(void) {
 void page_free(void *page) {
   /* Rudimentary double-free / out-of-range guard */
   uintptr_t addr = (uintptr_t)page;
-  uintptr_t pb = page_pool_base();
+  uint32_t pb = page_pool_base();
   if (addr < pb || addr >= pb + page_count * PAGE_SIZE)
     return; /* ignore bogus pointer rather than corrupt the stack */
 
@@ -333,9 +333,9 @@ void page_free(void *page) {
   spin_unlock_irqrestore(SPIN_PAGE, saved);
 }
 
-uintptr_t page_pool_base(void) {
+uint32_t page_pool_base(void) {
 #if defined(__xtensa__)
-  return xtensa_pool_base;
+  return (uint32_t)xtensa_pool_base;
 #else
   return PAGE_POOL_BASE;
 #endif
